# server/openapi.yaml
openapi: 3.1.0
info:
  title: Concord Cognitive Engine API
  description: |
    A local-first cognitive operating system built around DTUs (Discrete Thought Units).

    ## Authentication
    - **Cookie Auth**: httpOnly cookies (recommended for browsers)
    - **Bearer Token**: JWT in Authorization header
    - **API Key**: X-API-Key header for programmatic access

    ## Key Concepts
    - **DTU**: Discrete Thought Unit - the atomic unit of knowledge
    - **MEGA/HYPER**: Higher-order DTUs consolidating multiple DTUs
    - **Macros**: Deterministic functions for DTU operations
    - **Council**: Governance system for credibility and approval
  version: 5.1.0
  contact:
    name: Concord Team
  license:
    name: MIT
servers:
  - url: http://localhost:5050
    description: Local development
  - url: https://api.concord.example.com
    description: Production
...
paths:
...
  /api/dtus:
    get:
      tags: [DTUs]
      summary: List DTUs
      description: Returns all DTUs with optional filtering
      parameters:
        - name: tier
          in: query
          schema:
            type: string
            enum: [regular, mega, hyper, global]
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: DTU list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  dtus:
                    type: array
                    items:
                      $ref: '#/components/schemas/DTU'
    post:
      tags: [DTUs]
      summary: Create DTU
      description: Creates a new DTU directly
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DTUCreate'
      responses:
        '201':
          description: DTU created
...
  /api/ingest/url:
    post:
      tags: [Ingest]
      summary: Ingest from URL
      description: Fetch and ingest content from a URL into DTUs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Content ingested
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  dtusCreated:
                    type: integer
  /api/ingest/text:
    post:
      tags: [Ingest]
      summary: Ingest text content
      description: Ingest raw text content and create DTUs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                title:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Text ingested
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  dtusCreated:
                    type: integer
  /api/ingest:
    post:
      tags: [Ingest]
      summary: Unified ingest
      description: Unified ingestion endpoint accepting text or URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                url:
                  type: string
                  format: uri
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Content ingested
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  dtusCreated:
                    type: integer
  /api/ingest/queue:
    post:
      tags: [Ingest]
      summary: Queue for async ingest
      description: Queue content for asynchronous background ingestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                url:
                  type: string
                  format: uri
                tags:
                  type: array
                  items:
                    type: string
                priority:
                  type: string
                  enum: [low, normal, high]
                  default: normal
      responses:
        '202':
          description: Ingestion queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  jobId:
                    type: string

  # --- Verify ---
  /api/verify/feasibility:
    post:
      tags: [Verify]
      summary: Check DTU feasibility
      description: Evaluate whether a DTU is feasible and well-formed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dtuId]
              properties:
                dtuId:
                  type: string
      responses:
        '200':
          description: Feasibility result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  feasible:
                    type: boolean
                  score:
                    type: number
                  issues:
                    type: array
                    items:
                      type: string
  /api/verify/designScore:
    post:
      tags: [Verify]
      summary: Score DTU design
      description: Score the design quality of a DTU
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dtuId]
              properties:
                dtuId:
                  type: string
      responses:
        '200':
          description: Design score
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  score:
                    type: number
                  breakdown:
                    type: object
  /api/verify/conflictCheck:
    post:
      tags: [Verify]
      summary: Check for conflicts
      description: Check whether a DTU conflicts with existing knowledge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dtuId]
              properties:
                dtuId:
                  type: string
      responses:
        '200':
          description: Conflict check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  hasConflicts:
                    type: boolean
                  conflicts:
                    type: array
                    items:
                      type: object
  /api/verify/stressTest:
    post:
      tags: [Verify]
      summary: Stress test DTU
      description: Run stress tests against a DTU to find edge cases
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dtuId]
              properties:
                dtuId:
                  type: string
      responses:
        '200':
          description: Stress test results
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  passed:
                    type: boolean
                  results:
                    type: array
                    items:
                      type: object
  /api/verify/deriveSecondOrder:
    post:
      tags: [Verify]
      summary: Derive second-order effects
      description: Analyze a DTU for second-order implications and effects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dtuId]
              properties:
                dtuId:
                  type: string
      responses:
        '200':
          description: Second-order effects
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  effects:
                    type: array
                    items:
                      type: object
  /api/verify/lineageLink:
    post:
      tags: [Verify]
      summary: Link DTU lineage
      description: Establish lineage links between DTUs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceId, targetId]
              properties:
                sourceId:
                  type: string
                targetId:
                  type: string
                relation:
                  type: string
      responses:
        '200':
          description: Lineage linked
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean

  # --- Jobs ---
...
  /api/jobs/toggle:
    post:
      tags: [Jobs]
      summary: Toggle job enabled/disabled
      description: Enable or disable a specific background job type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [jobType, enabled]
              properties:
                jobType:
                  type: string
                enabled:
                  type: boolean
      responses:
        '200':
          description: Job toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  jobType:
                    type: string
                  enabled:
                    type: boolean
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: concord_auth
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    DTU:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        tier:
          type: string
          enum: [regular, mega, hyper]
        tags:
          type: array
          items:
            type: string
        core:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DTUCreate:
      type: object
      required: [content]
      properties:
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        source:
          type: string
        isGlobal:
          type: boolean
          default: false
...
security:
  - cookieAuth: []
  - bearerAuth: []
  - apiKeyAuth: []